[Floating-point] Changed `DTYPE_INT` to `int64` instead of `int32`

+++ src/jax_fdm/__init__.py
@@ -41,8 +41,8 @@ __all__ = ["HOME", "DATA", "DOCS", "TEMP"]
 DTYPE_NP = np.float64
 DTYPE_JAX = jnp.float64
 
-DTYPE_INT_NP = jnp.int32
-DTYPE_INT_JAX = jnp.int32
+DTYPE_INT_NP = jnp.int64
+DTYPE_INT_JAX = jnp.int64
 
 # this only works on startup!
 if DTYPE_JAX == jnp.float64 or DTYPE_NP == np.float64:
